Score the PID-5-FSF Validity Scales

Create a data frame with scores on the PID-5 Faceted Short Form validity scales.

Usage

validity_pid5fsf(
  data,
  items = NULL,
  id = NULL,
  scales = c("PNA", "INCS", "ORSS", "PRDS", "SDTDS"),
  srange = c(0, 3),
  tibble = FALSE
)

Arguments

data

A data frame containing all PID-5 items (numerically scored).

items

An optional vector of column names (as strings) or numbers (as integers) corresponding to the 100 PID-5 items in order. If set to NULL (the default), all non-id columns will be assumed to be the items in order.

id

An optional vector of column names (as strings) or numbers (as integers) corresponding to variables from data to keep in the output. If set to NULL (the default), no columns will be retained.

scales

An optional character vector indicating which of the following validity scales to calculate: percent of missing items ("PNA"), response inconsistency scale short form ("INCS"), over-reporting scale short form ("ORSS"), positive impression management response distortion scale short form ("PRDS"), and social desirability-total denial short form ("SDTDS"). See details below for interpretation guidance

srange

An optional numeric vector specifying the minimum and maximum values of the PID-5 items, used for reverse-coding. (default = c(0, 3))

tibble

An optional logical indicating whether the output should be converted to a tibble::tibble().

Value

A data frame containing any id variables as well any requested scale scores.

Details

Scores of 8 or more on the INC-S are indicative of inconsistent responding. Cut-scores for the ORS-S, PRD-S, and SD-TD-S have not yet been validated.

References

Lowmaster, S. E., Hartman, M. J., Zimmermann, J., Baldock, Z. C., & Kurtz, J. E. (2020). Further Validation of the Response Inconsistency Scale for the Personality Inventory for DSM-5. Journal of Personality Assessment, 102(6), 743–750. https://doi.org/10.1080/00223891.2019.1674320