Simulated HiTOP-SR Data

Simulated responses to items on the full HiTOP-SR (with 405 items). Note that this is a naive simulation where response options 1 to 4 are all equally likely and generated independently per item. Thus, responses are not clustered within scales, and these data can be used (eventually) to test validity tools intended to detect inconsistent/random responding.

Usage

sim_hitopsr

Format

A tibble with 100 rows and 405 columns.

hsr_1 to hsr_405

Responses on each item

Examples

sim_hitopsr
#> # A tibble: 100 × 405
#>    hsr_1 hsr_2 hsr_3 hsr_4 hsr_5 hsr_6 hsr_7 hsr_8 hsr_9 hsr_10 hsr_11 hsr_12
#>    <int> <int> <int> <int> <int> <int> <int> <int> <int>  <int>  <int>  <int>
#>  1     2     2     2     3     2     1     3     1     1      3      3      3
#>  2     3     4     4     1     1     3     4     1     1      3      4      4
#>  3     3     2     2     3     2     3     3     3     1      4      2      1
#>  4     3     3     4     2     2     3     4     4     1      1      1      2
#>  5     1     3     4     1     2     1     4     3     2      3      3      1
#>  6     1     1     2     3     1     2     2     4     1      3      2      4
#>  7     2     1     1     3     2     1     2     3     4      4      4      1
#>  8     1     1     3     3     1     1     3     2     3      3      3      2
#>  9     3     4     4     3     4     3     3     2     3      1      4      2
#> 10     2     4     1     1     2     2     3     2     2      3      1      4
#> # ℹ 90 more rows
#> # ℹ 393 more variables: hsr_13 <int>, hsr_14 <int>, hsr_15 <int>, hsr_16 <int>,
#> #   hsr_17 <int>, hsr_18 <int>, hsr_19 <int>, hsr_20 <int>, hsr_21 <int>,
#> #   hsr_22 <int>, hsr_23 <int>, hsr_24 <int>, hsr_25 <int>, hsr_26 <int>,
#> #   hsr_27 <int>, hsr_28 <int>, hsr_29 <int>, hsr_30 <int>, hsr_31 <int>,
#> #   hsr_32 <int>, hsr_33 <int>, hsr_34 <int>, hsr_35 <int>, hsr_36 <int>,
#> #   hsr_37 <int>, hsr_38 <int>, hsr_39 <int>, hsr_40 <int>, hsr_41 <int>, …