Score the Full PID-5 Validity Scales

Create a data frame with scores on the full PID-5 validity scales.

Usage

validity_pid5(
  data,
  items,
  scales = c("PNA", "INC", "ORS", "PRD", "SDTD"),
  srange = c(0, 3),
  append = TRUE,
  tibble = TRUE
)

Arguments

items

An optional vector of column names (as strings) or numbers (as integers) corresponding to the 220 PID-5 items in order. If set to NULL (the default), all non-id columns will be assumed to be the items in order.

scales

An optional character vector indicating which of the following validity scales to calculate: percent missing items ("PNA"), response inconsistency scale ("INC"), overreporting scale ("ORS"), positive impression management response distortion scale ("PRD"), and social desirability-total denial scale ("SDTD"). See details below for interpretation guidance.

srange

An optional numeric vector specifying the minimum and maximum values of the PID-5 items, used for reverse-coding. (default = c(0, 3))

tibble

An optional logical indicating whether the output should be converted to a tibble::tibble().

.data

A data frame containing all PID-5 items (numerically scored).

id

An optional vector of column names (as strings) or numbers (as integers) corresponding to variables from .data to keep in the output. If set to NULL (the default), no columns will be retained.

Value

A data frame containing any id variables as well any requested scale scores.

Details

A score of 17 or higher on the INC is indicative of inconsistent responding. A score of 3 or higher on the ORS is indicative of overreporting. A score of 10 or lower on the PRD is indicative of positive impression management (underreporting rather than genuine responding) whereas a score of 21 or higher on the PRD is indicative of genuine responding (rather than positive impression management). A score of 11 or lower on the SDTD is indicative of social desirability (rather than defensiveness), whereas a score of 19 or higher on the SDTD is indicative of defensiveness (rather than social desirability).

References

Keeley, J. W., Webb, C., Peterson, D., Roussin, L., & Flanagan, E. H. (2016). Development of a Response Inconsistency Scale for the Personality Inventory for DSM-5. Journal of Personality Assessment, 98(4), 351-359. https://doi.org/10.1080/00223891.2016.1158719

Sellbom, M., Dhillon, S., & Bagby, R. M. (2018). Development and validation of an overreporting scale for the Personality Inventory for DSM-5 (PID-5). Psychological Assessment, 30(5), 582-593. https://doi.org/10.1037/pas0000507

Bagby, R. M., & Sellbom, M. (2018). The Validity and Clinical Utility of the Personality Inventory for DSM–5 Response Inconsistency Scale. Journal of Personality Assessment, 100(4), 398–405. https://doi.org/10.1080/00223891.2017.1420659

Williams, M. M., Rogers, R., Sharf, A. J., & Ross, C. A. (2019). Faking Good: An Investigation of Social Desirability and Defensiveness in an Inpatient Sample With Personality Disorder Traits. Journal of Personality Assessment, 101(3), 253–263. https://doi.org/10.1080/00223891.2018.1455691